requirements/prod.txt
# Everything needed in production

# Flask
Flask==0.10.1
MarkupSafe==0.23
Werkzeug==0.9.6
Jinja2==2.7.3
itsdangerous==0.24

# Database
Flask-SQLAlchemy==2.0
SQLAlchemy==0.9.8

# Migrations
Flask-Migrate==1.3.0

# Forms
Flask-WTF==0.10.3
WTForms==2.0.1

# Deployment
waitress==0.8.9

# Assets
Flask-Assets==0.10
cssmin>=0.2.0
jsmin>=2.0.11

# Auth
Flask-Login==0.2.11
Flask-Bcrypt==0.6.0

# Caching
Flask-Cache>=0.13.1

# Debug toolbar
Flask-DebugToolbar==0.9.1

# Spellchecking
pyenchant==1.6.6
